move views to separate module views.py
